Swag on

+++ SmartImage.Lib/Clients/AnilistClient.cs
@@ -18,7 +18,7 @@ public sealed class AnilistClient : IDisposable
 		m_client = new GraphQLClient("https://graphql.anilist.co");
 	}
 
-	public async Task<string> GetTitleAsync(int anilistId)
+	public async Task<string> GetTitleAsync(long anilistId)
 	{
 		/*
 		 * https://anilist.gitbook.io/anilist-apiv2-docs/overview/graphql
